establish working definition of group
These will be the working definitions and notations that we will use
for the entirety of the project. Chapter 1 should build up to these
definitions, possibly using definitions of their own, while chapters 2
and onward should use these definitions in future theorems.

The namespace in which these definitions are located is meant to be
temporary, as it will be moved into a chapter once the structure of the
project becomes more concrete.