Added one example to Sheet0 based on previous
This example added uses the have tactic. Went back to Chapter1 to look
at the most commonly used tactics, and everything else was covered in
this interlude / refresher except for `have`, so wanted to make sure
that was in there as well.

Ready to move on to Morphisms in Sheet1.