made some variables implicit definition of Subgroup, updated the proof of subgroup lattice
In the mul_closure and inv_closure properties of a Subgroup, I made the variables a and b implicit to make the proofs shorter. I amended all the proofs that depend on this. In the proof that Subgroups form a partial order with the relation Subset, I removed replaced H.carrier with (H : Set G), which utilizes the coercion of a Subgroup G to a Set G. This is cleaner because the user will never need to use or see the word carrier. For instance, they will see $a \in Normalizer H$ instead of $a \in (Normalizer H).carrier$. No proofs change because of this.