Modifications to Sheet0, Almost done Sheet3
Sheet3 will be finished up once the last proof is completed (halfway
done as of writing this). Proof that conjugation under an arbitrary
group operation is an automorphism. Completed proof of homomorphism
property for this example, need proof of bijectivity still.

Should have automorphisms complete tomorrow, ready to merge with rest
of Chapter 2, after which blueprint will be installed as well, then
a pending PR with master.