remove bad simp lemmas
The RHS of a simp lemma should always be "simpler" than the LHS.
Commutativity is especially problematic with this, since there is no
information on which side is simpler. Since Basic.lean will be removed
when chapters are merged, this is more of a reminder to do it then
(since it should cause merge conflicts).