add mk_all to workflows
`lake build` will only build files that AlgebraInLean.lean
(transitively) depends upon, so if new files are missing, then that
check is useless.

The non-strict build only updates it within the workflow since the user
will presumably get around to it eventually. The strict build fails if
it needs to update it since it should always be correct in master.