Finished formatting for `Chapter02` - WOOOOOOOOOO
Formatting officially finished for all of Chapter 02 - Morphisms.
Indenting was fixed, as was all namespacing and import issues. Content
is finished and sample solutions are noted for deletion by the program
with `-- SAMPLE SOLUTION` and `-- END SAMPLE SOLUTION`, as was the
convention used in previous chapters. All formatting matches previous
chapters as well.

THIS LOOKS GOOOOOOOOOOOD.